Compute the expected values of different strategries deterministically
Previously, the expected outcomes of the different strategies have been
computed by a Monte-Carlo simulation, but this takes many iterations to
collect enough data points to resemble the random distribution.
The computation blocks the event queue, which results in delayed and
stuttering animations at the end of each game.

This patch replaces the Monte-Carlo simulation with a deterministic
approach whose complexity is at the order of a single Monte-Carlo
iteration.

As a side effect, the result is also more precise since the Monte-Carlo
method only approaches the exact result in the limit when the number of
iterations grows towards infinity.